backend/requirements.txt
fastapi
uvicorn
scikit-learn
Pillow
pytesseract
python-multipart
bcrypt
python-jose[cryptography]
motor
pydantic[email]
